refactor : 캐러셀 wrapper 분리
