+++ src/pages/Homepage/components/Carousel/Carousel.jsx
@@ -2,38 +2,13 @@ import { useCallback, useEffect, useRef } from 'react';
 import { NextButton, PrevButton, usePrevNextButtons } from './EmblaCarouselArrowButtons';
 import { Link } from 'react-router';
 import useEmblaCarousel from 'embla-carousel-react';
-import PNG_IMAGES from '@public/images/image';
+import PropTypes from 'prop-types';
 import './style.css';
 
 const TWEEN_FACTOR_BASE = 0.84;
 const numberWithinRange = (number, min, max) => Math.min(Math.max(number, min), max);
-// TODO wrapper로 분리
-const items = [
-  {
-    title: '제주속 제주 휴애리 겨울 동백과 함께하는 여행',
-    subtitle: 'T도 F만들어 버리는 아름다운 동백꽃 풍경',
-    imageName: '서귀포시 자연생활공원',
-    imageURL: PNG_IMAGES.carousel_1,
-    contentId: 'CONT_000000000500707',
-  },
-  {
-    title: '성산일출봉이 한눈에 보이는 제주도 대표 일출명소',
-    subtitle: '<웰컴투 삼달리> 드라마속 제주를 찾는 여정',
-    imageName: '서귀포시 오조포구',
-    imageURL: PNG_IMAGES.carousel_2,
-    contentId: 'CNTS_000000000020987',
-  },
-  {
-    title: '메밀꽃과 라벤더를 한아름 품은 농장',
-    subtitle: '꽃과 동물들에 둘러쌓인 힐링의 명소',
-    imageName: '서귀포시 포선면 보롬왓',
-    imageURL: PNG_IMAGES.carousel_3,
-    contentId: 'CNTS_000000000020050',
-  },
-];
 
-const Carousel = () => {
-  const options = { align: 'center', loop: true };
+const Carousel = ({ options, items }) => {
   const [emblaRef, emblaApi] = useEmblaCarousel(options);
   const tweenFactor = useRef(0);
   const { prevBtnDisabled, nextBtnDisabled, onPrevButtonClick, onNextButtonClick } =
@@ -127,3 +102,19 @@ const Carousel = () => {
   );
 };
 export default Carousel;
+
+Carousel.propTypes = {
+  options: PropTypes.shape({
+    align: PropTypes.string.isRequired,
+    loop: PropTypes.bool.isRequired,
+  }),
+  items: PropTypes.arrayOf(
+    PropTypes.shape({
+      title: PropTypes.string.isRequired,
+      subtitle: PropTypes.string.isRequired,
+      imageName: PropTypes.string.isRequired,
+      imageURL: PropTypes.string.isRequired,
+      contentId: PropTypes.string.isRequired,
+    }),
+  ),
+};
