+++ src/pages/Homepage/components/Carousel/CarouselWrapper.jsx
@@ -1,0 +1,33 @@
+import Carousel from './Carousel';
+import PNG_IMAGES from '@public/images/image';
+
+const CarouselWrapper = () => {
+  const items = [
+    {
+      title: '제주속 제주 휴애리 겨울 동백과 함께하는 여행',
+      subtitle: 'T도 F만들어 버리는 아름다운 동백꽃 풍경',
+      imageName: '서귀포시 자연생활공원',
+      imageURL: PNG_IMAGES.carousel_1,
+      contentId: 'CONT_000000000500707',
+    },
+    {
+      title: '성산일출봉이 한눈에 보이는 제주도 대표 일출명소',
+      subtitle: '<웰컴투 삼달리> 드라마속 제주를 찾는 여정',
+      imageName: '서귀포시 오조포구',
+      imageURL: PNG_IMAGES.carousel_2,
+      contentId: 'CNTS_000000000020987',
+    },
+    {
+      title: '메밀꽃과 라벤더를 한아름 품은 농장',
+      subtitle: '꽃과 동물들에 둘러쌓인 힐링의 명소',
+      imageName: '서귀포시 포선면 보롬왓',
+      imageURL: PNG_IMAGES.carousel_3,
+      contentId: 'CNTS_000000000020050',
+    },
+  ];
+
+  const options = { align: 'center', loop: true };
+  return <Carousel options={options} items={items} />;
+};
+
+export default CarouselWrapper;
