+++ src/pages/Homepage/components/index.js
@@ -1,2 +1,2 @@
 export { default as Hero } from './Hero';
-export { default as Carousel } from './Carousel';
+export { default as CarouselWrapper } from './Carousel/CarouselWrapper';
